Moved folder as an util, added namespace
