+++ src/ts/vttutils.ts
@@ -1,4 +1,4 @@
-import { SubtitleRegionContainer, SubtitleLabel } from './subtitleoverlay';
+import { SubtitleRegionContainer, SubtitleLabel } from './components/subtitleoverlay';
 import { VTTProperties, VTTRegionProperties } from 'bitmovin-player/types/subtitles/vtt/API';
 
 // Our default height of a line
@@ -39,7 +39,7 @@ const setDefaultVttStyles = (cueContainer: SubtitleLabel) => {
  * Align the Cue Box's line
  * https://w3.org/TR/webvtt1/#webvtt-cue-line-alignment
  */
-const setVttLineAlign = (cueContainer: SubtitleLabel, {lineAlign}: VTTProperties, direction: Direction) => {
+const setVttLineAlign = (cueContainer: SubtitleLabel, { lineAlign }: VTTProperties, direction: Direction) => {
   switch (lineAlign) {
     case 'center':
       cueContainer.getDomElement().css(`margin-${direction}`, `${-lineHeight / 2}px`);
@@ -50,7 +50,7 @@ const setVttLineAlign = (cueContainer: SubtitleLabel, {lineAlign}: VTTProperties
 };
 
 /**
- * Defines the positioning of the Cue Box
+ * Defines the line positioning of the Cue Box
  * https://w3.org/TR/webvtt1/#webvtt-cue-line
  */
 const setVttLine = (cueContainer: SubtitleLabel, vtt: VTTProperties, direction: Direction) => {
@@ -58,17 +58,17 @@ const setVttLine = (cueContainer: SubtitleLabel, vtt: VTTProperties, direction: 
     if (!vtt.snapToLines) {
       cueContainer.getDomElement().css(direction, vtt.line as string);
       cueContainer.getDomElement().css(DirectionPair.get(direction), 'unset');
-  
+
       setVttLineAlign(cueContainer, vtt, direction);
     } else if (vtt.snapToLines && vtt.line > 0) {
       cueContainer.getDomElement().css(direction, `${vtt.line as number * lineHeight}px`);
       cueContainer.getDomElement().css(DirectionPair.get(direction), 'unset');
-  
+
       setVttLineAlign(cueContainer, vtt, direction);
     } else if (vtt.snapToLines && vtt.line < 0) {
       cueContainer.getDomElement().css(DirectionPair.get(direction), `${vtt.line as number * -lineHeight}px`);
       cueContainer.getDomElement().css(direction, 'unset');
-  
+
       setVttLineAlign(cueContainer, vtt, DirectionPair.get(direction));
     }
   }
@@ -81,8 +81,6 @@ const setVttLine = (cueContainer: SubtitleLabel, vtt: VTTProperties, direction: 
 const setVttWritingDirection = (cueContainer: SubtitleLabel, vtt: VTTProperties) => {
   if (vtt.vertical === '') {
     cueContainer.getDomElement().css('writing-mode', 'horizontal-tb');
-    cueContainer.getDomElement().css('left', '0');
-    cueContainer.getDomElement().css('bottom', '0');
     setVttLine(cueContainer, vtt, Direction.Top);
   } else if (vtt.vertical === 'lr') {
     cueContainer.getDomElement().css('writing-mode', 'vertical-lr');
@@ -121,37 +119,46 @@ const setVttPositionAlign = (cueContainer: SubtitleLabel, vtt: VTTProperties, di
   }
 };
 
-export const setVttCueBoxStyles = (cueContainer: SubtitleLabel, vtt: VTTProperties) => {
-  console.warn(vtt);
-  setDefaultVttStyles(cueContainer);
+export namespace VttUtils {
+  export const setVttCueBoxStyles = (cueContainer: SubtitleLabel, vtt: VTTProperties) => {
+    setDefaultVttStyles(cueContainer);
 
-  setVttWritingDirection(cueContainer, vtt);
+    setVttWritingDirection(cueContainer, vtt);
 
-  // https://w3.org/TR/webvtt1/#webvtt-cue-text-alignment
-  const textAlign = vtt.align === 'middle' ? 'center' : vtt.align;
-  cueContainer.getDomElement().css('text-align', textAlign);
+    // https://w3.org/TR/webvtt1/#webvtt-cue-text-alignment
+    const textAlign = vtt.align === 'middle' ? 'center' : vtt.align;
+    cueContainer.getDomElement().css('text-align', textAlign);
 
-  const containerSize = vtt.size;
-  // https://w3.org/TR/webvtt1/#webvtt-cue-size
-  if (vtt.vertical === '') {
-    cueContainer.getDomElement().css('width', `${containerSize}%`);
-    setVttPositionAlign(cueContainer, vtt, Direction.Left);
-  } else {
-    cueContainer.getDomElement().css('height', `${containerSize}%`);
-    setVttPositionAlign(cueContainer, vtt, Direction.Top);
-  }
-};
+    // https://www.w3.org/TR/webvtt1/#webvtt-cue-position
+    if (vtt.position === 'auto') {
+      cueContainer.getDomElement().css('left', '0');
+      cueContainer.getDomElement().css('bottom', '0');
+    }
 
-// https://www.w3.org/TR/webvtt1/#regions
-export const setVttRegionStyles = (regionContainer: SubtitleRegionContainer, region: VTTRegionProperties, overlaySize: {width: number, height: number}) => {
-  const regionPositionX = overlaySize.width * region.viewportAnchorX / 100 - ((overlaySize.width * region.width / 100) * region.regionAnchorX / 100);
-  const regionPositionY = overlaySize.height * region.viewportAnchorY / 100 - ((region.lines * lineHeight) * region.regionAnchorY / 100);
-  regionContainer.getDomElement().css('position', 'absolute');
-  regionContainer.getDomElement().css('overflow', 'hidden');
-  regionContainer.getDomElement().css('width', `${region && region.width ? region.width : 100}%`);
-  regionContainer.getDomElement().css('left', `${regionPositionX}px`);
-  regionContainer.getDomElement().css('right', 'unset');
-  regionContainer.getDomElement().css('top', `${regionPositionY}px`);
-  regionContainer.getDomElement().css('bottom', 'unset');
-  regionContainer.getDomElement().css('height', `${region && region.lines ? region.lines * lineHeight : 3 * lineHeight}px`);
-};
+    // https://w3.org/TR/webvtt1/#webvtt-cue-size
+    const containerSize = vtt.size;
+    if (vtt.vertical === '') {
+      cueContainer.getDomElement().css('width', `${containerSize}%`);
+      setVttPositionAlign(cueContainer, vtt, Direction.Left);
+    } else {
+      cueContainer.getDomElement().css('height', `${containerSize}%`);
+      setVttPositionAlign(cueContainer, vtt, Direction.Top);
+    }
+  };
+
+  /** https://www.w3.org/TR/webvtt1/#regions
+   *  https://www.speechpad.com/captions/webvtt#toc_16
+   */
+  export const setVttRegionStyles = (regionContainer: SubtitleRegionContainer, region: VTTRegionProperties, overlaySize: { width: number, height: number }) => {
+    const regionPositionX = overlaySize.width * region.viewportAnchorX / 100 - ((overlaySize.width * region.width / 100) * region.regionAnchorX / 100);
+    const regionPositionY = overlaySize.height * region.viewportAnchorY / 100 - ((region.lines * lineHeight) * region.regionAnchorY / 100);
+    regionContainer.getDomElement().css('position', 'absolute');
+    regionContainer.getDomElement().css('overflow', 'hidden');
+    regionContainer.getDomElement().css('width', `${region && region.width ? region.width : 100}%`);
+    regionContainer.getDomElement().css('left', `${regionPositionX}px`);
+    regionContainer.getDomElement().css('right', 'unset');
+    regionContainer.getDomElement().css('top', `${regionPositionY}px`);
+    regionContainer.getDomElement().css('bottom', 'unset');
+    regionContainer.getDomElement().css('height', `${region && region.lines ? region.lines * lineHeight : 3 * lineHeight}px`);
+  };
+}
